Moved the "main" subproject logic underneath the airline project. Also removed "util" as a package name
